Refactor state loading and saving
We abstract the loading and saving of persistent state. We do this
so we can introduce e.g. S3 backed persistence in the future.

While we're here, we use codecs.open() to avoid extra encoding foo. We
also handle missing state files and return the default empty state
specified by the caller.